Move contents of `white` module into `stream` and `adapters`
